Add files via upload
Here is the current version of the code. Both letters and numbers have been added along with a home screen and the text to speech functionality. I am currently attempting to create a region around the number that would only allow for tracing within the boundaries of the number. After I get through with that (if I get through), I would attempt to track if the entire number has been traced to be accepted as valid input, If not entirely traced an error message would be displayed.